change long to double in metric tpct

+++ code/self_managment/src/main/java/com/self_managment/persistance/dao/SummaryDao.java
@@ -15,7 +15,7 @@ public interface SummaryDao extends GenericDao<Summary, Serializable> {
     public Long getNCH(Integer campaignId, Integer supervisorId,
 	    Integer docket, Date dateFrom, Date dateTo);
 
-    public Long getTransferPCT(Integer campaignId, Integer supervisorId,
+    public Double getTransferPCT(Integer campaignId, Integer supervisorId,
 	    Integer docket, Date dateFrom, Date dateTo);
 
     public Long getTotalLoggedTime(Integer campaignId, Integer supervisorId,
